fix: change Long on Integer in method getMovieRecommendations

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -94,8 +94,8 @@ public class UserService {
                 .collect(Collectors.toList());
     }
 
-    public List<Film> getMovieRecommendations(Long userId) {
-        getById(userId);
+    public List<Film> getMovieRecommendations(Integer userId) {
+        getById(Long.valueOf(userId));
         return filmStorage.getMovieRecommendations(userId);
     }
