+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/db/DbFilmStorage.java
@@ -158,7 +158,7 @@ public class DbFilmStorage extends DbStorage implements Storage<Film> {
         return new ArrayList<>(films.values());
     }
 
-    public List<Film> getMovieRecommendations(Long userId) {
+    public List<Film> getMovieRecommendations(Integer userId) {
         String sql = "" +
                 "SELECT tb1.id, tb1.name, tb1.description, tb1.releasedate, tb1.duration, tb1.mpa_id, m.name as mpa_name" +
                 ", fg.genres_id, g.name as genres_name, tb1.users_id " +
